BaytesDiff

Documentation for BaytesDiff.

BaytesDiff.AbstractDiffOrder

BaytesDiff.AbstractDifferentiableTune

BaytesDiff.AnalyticalDiffTune

BaytesDiff.AutomaticDifferentiationMethod

BaytesDiff.DiffObjective

BaytesDiff.ObjectiveError

BaytesDiff.ℓDensityResult

BaytesDiff.ℓGradientResult

BaytesDiff.ℓHessianResult

BaytesDiff.ℓObjectiveResult

BaytesDiff._config

BaytesDiff._diffresults_gradientbuffer

BaytesDiff._diffresults_hessianbuffer

BaytesDiff.check_gradients

BaytesDiff.log_density

BaytesDiff.log_density_and_gradient

BaytesDiff.log_density_and_gradient_and_hessian

BaytesDiff.AbstractDiffOrder — Type

abstract type AbstractDiffOrder

Abstract super type that determines differentiability of objective function and configuration that is created when initiating a AutomaticDiffTune struct.

BaytesDiff.AbstractDifferentiableTune — Type

abstract type AbstractDifferentiableTune

Abstract super type for Tuning structs of differentiable functions.

BaytesDiff.AnalyticalDiffTune — Type

struct AnalyticalDiffTune{G<:Function, H<:Union{Nothing, Function}} <: BaytesDiff.AbstractDifferentiableTune

Stores information for evaluating and taking the gradient of an objective function.

Fields

gradient::Function: Gradient as function of ℓobjective and parameter vector in unconstrained space, gradient(ℓobjective, θᵤ).

hessian::Union{Nothing, Function}: Hessian as function of ℓobjective and parameter vector in unconstrained space, hessian(ℓobjective, θᵤ).

BaytesDiff.AutomaticDifferentiationMethod — Type

abstract type AutomaticDifferentiationMethod

Abstract super type for Supported Automatic Differentiation backends.

BaytesDiff.DiffObjective — Type

struct DiffObjective{O<:ModelWrappers.Objective, T<:BaytesDiff.AbstractDifferentiableTune}

Objective struct with additional information about AD backend and configuration.

Fields

objective::ModelWrappers.Objective: Objective as function of a parameter vector in unconstrained space.

tune::BaytesDiff.AbstractDifferentiableTune: Automatic Differentiation configurations.

BaytesDiff.ObjectiveError — Type

struct ObjectiveError <: Exception

Stores parameter in constrained space at which logdensity could not be evaluated.

Fields

msg::String

ℓθᵤ::Real

θ::NamedTuple

θᵤ::AbstractVector

BaytesDiff.ℓDensityResult — Type

struct ℓDensityResult{T, S} <: ℓObjectiveResult

Stores result for log density and parameter for 'ℓobjective' evaluation at 'parameter'.

Fields

θᵤ::Any: Parameter in unconstrained space.

ℓθᵤ::Any: Log density at θᵤ.

BaytesDiff.ℓGradientResult — Type

struct ℓGradientResult{T<:(AbstractVector), S<:Real, G<:(AbstractVector)} <: ℓObjectiveResult

Stores result for log density, gradient, and parameter for 'ℓobjective' evaluation at 'parameter'.

Fields

θᵤ::AbstractVector: Parameter in unconstrained space.

ℓθᵤ::Real: Log density at θᵤ.

∇ℓθᵤ::AbstractVector: Gradient of log density at θᵤ.

BaytesDiff.ℓHessianResult — Type

struct ℓHessianResult{T<:(AbstractVector), S<:Real, G<:(AbstractVector), H<:(AbstractMatrix)} <: ℓObjectiveResult

Stores result for log density, gradient, hessian and parameter for 'ℓobjective' evaluation at 'parameter'.

Fields

θᵤ::AbstractVector: Parameter in unconstrained space.

ℓθᵤ::Real: Log density at θᵤ.

∇ℓθᵤ::AbstractVector: Gradient of log density at θᵤ.

Δℓθᵤ::AbstractMatrix: Hessian of log density at θᵤ.

BaytesDiff.ℓObjectiveResult — Type

abstract type ℓObjectiveResult <: BaytesCore.AbstractResult

Abstract super type for AbstractDifferentiableObjective results.

BaytesDiff._config — Function

Write config file for AD wrapper.

Examples

BaytesDiff._diffresults_gradientbuffer — Method

_diffresults_gradientbuffer(θᵤ)

Initiate DiffResults.MutableDiffResult struct buffer for gradients. Not exported.

Examples

BaytesDiff._diffresults_hessianbuffer — Method

_diffresults_hessianbuffer(θᵤ)

Initiate DiffResults.MutableDiffResult struct buffer for hessian. Not exported.

Examples

BaytesDiff.check_gradients — Function

check_gradients(_rng, objective)
check_gradients(_rng, objective, ADlibraries)
check_gradients(_rng, objective, ADlibraries, θᵤ)
check_gradients(
    _rng,
    objective,
    ADlibraries,
    θᵤ,
    difftune;
    printoutput
)

Check gradient computations of different backends against 'objective'.

Examples

BaytesDiff.log_density — Function

log_density(objective, θᵤ)

Compute log density of 'objective' at 'θᵤ'.

Examples

BaytesDiff.log_density_and_gradient — Function

log_density_and_gradient(objective, θᵤ)

Compute log density and gradient of 'objective' at 'θᵤ'.

Examples

BaytesDiff.log_density_and_gradient_and_hessian — Function

log_density_and_gradient_and_hessian(objective, θᵤ)

Compute log density, gradient and hessian of 'objective' at 'θᵤ'.

Examples